+++ helpers/test/integration/proxy_io_test/proxy_io_test.py
@@ -25,7 +25,12 @@ def storage_id():
 
 
 @pytest.fixture
-def file_uuid():
+def parameters():
+    return random_params()
+
+
+@pytest.fixture
+def file_id():
     return random_str()
 
 
@@ -40,61 +45,79 @@ def helper(storage_id, endpoint):
                                  endpoint.port)
 
 
-def test_write_should_write_data(file_uuid, storage_id, endpoint, helper):
-    wrote = random_int()
-    file_id = random_str()
-    data = random_str()
-    offset = random_int()
+@pytest.fixture
+def file_ctx(helper, file_id, parameters, request):
+    ctx = helper.open(file_id, parameters)
+    request.addfinalizer(lambda: helper.release(ctx, file_id))
+    return ctx
 
+
+def remote_data_msg(data):
+    server_message = messages_pb2.ServerMessage()
+    server_message.proxyio_response.status.code = common_messages_pb2.Status.ok
+    server_message.proxyio_response.remote_data.data = data
+    return server_message
+
+
+def remote_write_result_msg(wrote):
     server_message = messages_pb2.ServerMessage()
     server_message.proxyio_response.status.code = common_messages_pb2.Status.ok
     server_message.proxyio_response.remote_write_result.wrote = wrote
+    return server_message
+
+
+def test_write_should_write_data(file_ctx, file_id, parameters, storage_id,
+                                 endpoint, helper):
+    data = random_str()
+    offset = random_int()
+    server_message = remote_write_result_msg(len(data))
 
     with reply(endpoint, server_message) as queue:
-        assert wrote == helper.write(file_id, data, offset, file_uuid)
+        assert len(data) == helper.write(file_ctx, file_id, data, offset)
         received = queue.get()
 
     assert received.HasField('proxyio_request')
 
     request = received.proxyio_request
-    assert request.file_uuid == file_uuid
+    assert decode_params(request.parameters) == parameters
     assert request.storage_id == storage_id
     assert request.file_id == file_id
 
     assert request.HasField('remote_write')
-    assert request.remote_write.offset == offset
-    assert request.remote_write.data == data
+    assert request.remote_write.byte_sequence[0].offset == offset
+    assert request.remote_write.byte_sequence[0].data == data
 
 
-def test_write_should_pass_errors(endpoint, helper):
+def test_close_should_pass_write_errors(file_id, endpoint, helper, parameters):
     server_message = messages_pb2.ServerMessage()
     server_message.proxyio_response.status.code = \
         common_messages_pb2.Status.eacces
 
+    ctx = helper.open(file_id, parameters)
+
+    with reply(endpoint, server_message):
+        helper.write(ctx, file_id, random_str(), random_int())
+
     with pytest.raises(RuntimeError) as excinfo:
-        with reply(endpoint, server_message):
-            helper.write(random_str(), random_str(), random_int(), random_str())
+        helper.release(ctx, file_id)
 
     assert 'Permission denied' in str(excinfo.value)
 
 
-def test_read_should_read_data(file_uuid, storage_id, endpoint, helper):
+def test_read_should_read_data(file_ctx, file_id, parameters, storage_id,
+                               endpoint, helper):
     data = random_str()
-    file_id = random_str()
     offset = random_int()
-
-    server_message = messages_pb2.ServerMessage()
-    server_message.proxyio_response.status.code = common_messages_pb2.Status.ok
-    server_message.proxyio_response.remote_data.data = data
+    server_message = remote_data_msg(data)
 
     with reply(endpoint, server_message) as queue:
-        assert data == helper.read(file_id, offset, len(data), file_uuid)
+        assert data == helper.read(file_ctx, file_id, offset, len(data))
         received = queue.get()
 
     assert received.HasField('proxyio_request')
 
     request = received.proxyio_request
-    assert request.file_uuid == file_uuid
+    assert decode_params(request.parameters) == parameters
     assert request.storage_id == storage_id
     assert request.file_id == file_id
 
@@ -103,13 +126,63 @@ def test_read_should_read_data(file_uuid, storage_id, endpoint, helper):
     assert request.remote_read.size == len(data)
 
 
-def test_read_should_pass_errors(endpoint, helper):
+def test_read_should_pass_errors(file_ctx, file_id, endpoint, helper,
+                                 parameters):
     server_message = messages_pb2.ServerMessage()
     server_message.proxyio_response.status.code = \
         common_messages_pb2.Status.eperm
 
     with pytest.raises(RuntimeError) as excinfo:
         with reply(endpoint, server_message):
-            helper.read(random_str(), random_int(), random_int(), random_str())
+            helper.read(file_ctx, file_id, random_int(), random_int())
 
     assert 'Operation not permitted' in str(excinfo.value)
+
+
+def test_should_cache_read(file_ctx, file_id, parameters, storage_id, endpoint,
+                           helper):
+    data = random_str()
+    offset = random_int()
+    server_message = remote_data_msg(data)
+
+    with reply(endpoint, server_message):
+        assert data == helper.read(file_ctx, file_id, offset, len(data))
+
+    # Doesn't need a second reply from the provider
+    assert data == helper.read(file_ctx, file_id, offset, len(data))
+
+
+def test_should_invalidate_read_cache_on_write(file_ctx, file_id, parameters,
+                                               storage_id, endpoint, helper):
+    data = random_str()
+    offset = random_int()
+    server_message = remote_data_msg(data)
+
+    with reply(endpoint, server_message):
+        assert data == helper.read(file_ctx, file_id, offset, len(data))
+        # At this point the data will be cached
+
+    new_data = random_str()
+    server_message = remote_write_result_msg(len(new_data))
+
+    with reply(endpoint, server_message):
+        helper.write(file_ctx, file_id, new_data, offset)
+
+    server_message = remote_data_msg(new_data)
+    with reply(endpoint, server_message):
+        assert new_data == helper.read(file_ctx, file_id, offset, len(new_data))
+
+
+def test_should_flush_writes_on_close(file_id, parameters, storage_id, endpoint,
+                                      helper):
+    data = random_str()
+    offset = random_int()
+    server_message = remote_write_result_msg(len(data))
+
+    ctx = helper.open(file_id, parameters)
+
+    helper.write(ctx, file_id, data, offset)
+    assert 0 == endpoint.all_messages_count()
+
+    with reply(endpoint, server_message) as queue:
+        helper.release(ctx, file_id)