+++ src/cache/metadataCache.cc
@@ -7,12 +7,13 @@
  */
 
 #include "metadataCache.h"
+#include "fuseOperations.h"
 
 #include "logging.h"
-#include "scheduler.h"
 #include "messages/fuse/getFileAttr.h"
 #include "messages/fuse/getFileLocation.h"
 #include "messages/fuse/rename.h"
+#include "scheduler.h"
 
 #include <chrono>
 
@@ -79,7 +80,7 @@ void MetadataCache::getAttr(
 {
     if (!m_pathToUuid.insert(uuidAcc, path)) {
         getAttr(metaAcc, uuidAcc->second);
-        metaAcc->second.path = path;
+        metaAcc->second.paths.emplace(path);
         return;
     }
 
@@ -89,7 +90,7 @@ void MetadataCache::getAttr(
         m_metaCache.insert(metaAcc, attr.uuid());
         uuidAcc->second = attr.uuid();
         metaAcc->second.attr = std::move(attr);
-        metaAcc->second.path = path;
+        metaAcc->second.paths.emplace(path);
     }
     catch (...) {
         if (!metaAcc.empty())
@@ -157,14 +158,29 @@ void MetadataCache::rename(
         auto &uuid = metaAcc->second.attr.get().uuid();
 
         DLOG(INFO) << "Renaming file " << uuid << " to " << newPath;
-        auto future = m_communicator.communicate<messages::fuse::FuseResponse>(
-            messages::fuse::Rename{uuid, newPath});
+        auto paths = metaAcc->second.paths;
 
-        communication::wait(future);
+        if (metaAcc->second.removedUpstream) {
+            metaAcc->second.paths.erase(oldPath);
+        }
+        else {
+            auto future =
+                m_communicator.communicate<messages::fuse::FuseResponse>(
+                    messages::fuse::Rename{uuid, newPath});
 
-        metaAcc->second.path = newPath;
-        newUuidAcc->second = uuid;
+            communication::wait(future);
+            metaAcc->second.paths.clear();
+        }
+
+        metaAcc->second.paths.emplace(newPath);
         m_pathToUuid.erase(oldUuidAcc);
+        newUuidAcc->second = uuid;
+
+        if (!metaAcc->second.removedUpstream) {
+            metaAcc.release();
+            for (auto &path : paths)
+                m_pathToUuid.erase(path);
+        }
     }
     catch (...) {
         m_pathToUuid.erase(newUuidAcc);
@@ -181,7 +197,7 @@ void MetadataCache::map(Path path, std::string uuid)
     m_metaCache.insert(metaAcc, uuid);
 
     uuidAcc->second = std::move(uuid);
-    metaAcc->second.path = std::move(path);
+    metaAcc->second.paths.emplace(std::move(path));
 }
 
 void MetadataCache::map(Path path, FileLocation location)
@@ -193,20 +209,38 @@ void MetadataCache::map(Path path, FileLocation location)
     m_metaCache.insert(metaAcc, location.uuid());
 
     uuidAcc->second = location.uuid();
-    metaAcc->second.path = std::move(path);
+    metaAcc->second.paths.emplace(std::move(path));
     metaAcc->second.location = std::move(location);
 }
 
 void MetadataCache::remove(UuidAccessor &uuidAcc, MetaAccessor &metaAcc)
 {
+    auto paths = metaAcc->second.paths;
     m_metaCache.erase(metaAcc);
     m_pathToUuid.erase(uuidAcc);
+
+    for (auto &path : paths)
+        m_pathToUuid.erase(path);
+}
+
+void MetadataCache::removePathMappings(
+    UuidAccessor &uuidAcc, MetaAccessor &metaAcc)
+{
+    metaAcc->second.paths.erase(uuidAcc->first);
+    m_pathToUuid.erase(uuidAcc);
+
+    auto paths = metaAcc->second.paths;
+    metaAcc->second.paths.clear();
+    metaAcc.release();
+
+    for (auto &path : paths)
+        m_pathToUuid.erase(path);
 }
 
 void MetadataCache::removePathMapping(
     UuidAccessor &uuidAcc, MetaAccessor &metaAcc)
 {
-    metaAcc->second.path = boost::none;
+    metaAcc->second.paths.erase(uuidAcc->first);
     m_pathToUuid.erase(uuidAcc);
 }
 
@@ -216,15 +250,11 @@ void MetadataCache::remove(const std::string &uuid)
     if (!m_metaCache.find(metaAcc, uuid))
         return;
 
-    if (metaAcc->second.path) {
-        UuidAccessor uuidAcc;
-        if (m_pathToUuid.find(uuidAcc, metaAcc->second.path.get())) {
-            remove(uuidAcc, metaAcc);
-            return;
-        }
-    }
-
+    auto paths = metaAcc->second.paths;
     m_metaCache.erase(metaAcc);
+
+    for (auto &path : paths)
+        m_pathToUuid.erase(path);
 }
 
 std::size_t MetadataCache::PathHash::hash(const Path &path)
@@ -252,7 +282,16 @@ bool MetadataCache::waitForNewLocation(const std::string &uuid,
             location.blocks().end();
     };
 
-    return pair.second.wait_for(lock, timeout, pred);
+    for (auto t = 0ms; t < timeout; ++t) {
+        if (helpers::fuseInterrupted())
+            throw std::system_error{
+                std::make_error_code(std::errc::operation_canceled)};
+
+        if (pair.second.wait_for(lock, 1ms, pred))
+            return true;
+    }
+
+    return false;
 }
 
 void MetadataCache::notifyNewLocationArrived(const std::string &uuid)